requirements.txt
opencv-python>=4.8.0
ultralytics
torch
torchvision
albumentations
pyyaml